version_requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv
